requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==1.10.22
python-dotenv==1.0.0
typing-extensions==4.5.0

# AI and transcription services
openai==1.3.7
replicate==0.22.0
httpx==0.27.0

# Video and audio processing
moviepy==1.0.3
pydub==0.25.1
imageio-ffmpeg==0.4.9

# Web interface
jinja2==3.1.2
aiofiles==23.2.1

# Utilities
typing-extensions==4.8.0
uuid==1.30

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
ruff==0.1.6
mypy==1.7.1
black==23.11.0

# Optional dependencies for testing
requests==2.31.0